Finally implemented expanded view for gallery images via modal element. Added on-click functionality to total reviews to scroll to review component.
