+++ client/src/reducers/index.js
@@ -1,5 +1,6 @@
 import { combineReducers } from 'redux';
 
+import expandedViewReducer from './expandedViewReducer';
 import galleryReducer from './galleryReducer';
 import metaDataReducer from './metaDataReducer';
 import productInfoReducer from './productInfoReducer';
@@ -14,6 +15,7 @@ import reviewsReducer from './reviewsReducer';
 //IMPORT ALL REDUCERS AS THEY ARE CREATED AND ADD THEM TO THE ROOT REDUCER
 
 const reducers = {
+  expandedView: expandedViewReducer,
   metaData: metaDataReducer,
   qList: qListReducer,
   styles: stylesReducer,
